Implement full PID solution
