+++ src/ImageProcessor/config.ini
@@ -30,9 +30,13 @@ poll_interval_in_ms = 100
 
 [angle_controller]
 k_P = 1
+k_I = 0.001
+k_D = 1
 
 [distance_controller]
-k_p = 1
+k_P = 1
+k_I = 0.001
+k_D = 1
 margin = 10
 
 [other]
