+++ src/ImageProcessor/requirements.txt
@@ -3,3 +3,4 @@ numpy==1.19.4; platform_system == "Linux"
 opencv-python==4.4.0.46
 pyserial==3.5
 picamera==1.13; platform_system == "Linux"
+simple-pid==0.2.4
